Add files via upload
The output for the game page.


The output for the game page after giving the input


Click on the following link and play with your friend and you will understand the flow of the game. https://jynyhy9vu5r8xz5zy0utww-on.drv.tw/project_C89_C90_C91_C92.com/game_page.html

In this project, we will-
Write javascript code for the send() function, this function is called when send button is clicked from quiz_game_page.html
You need to write JS code in quiz_game_page.js, which you had created and gave a link in quiz_game_page.html like this -

In the previous project. Send function -

First, define the send() function


. Get the 2 numbers from the input box.

And do multiplication of this 2 number and store it inside a variable


parsInt() is used to convert a String to Integer.
Syntax of parsInt() - parsInt(variable) // This variable will contain a string.
String means it’s a sentence.
Integer means it’s a number.
When we take value from an input box it’s a string, so first we will convert this string to an integer using parsInt(). Then we will multiply these 2 numbers.
Now create the question div as per the image below -

So first create a heading tag with will hold “ Number1 X Number2”, and store it inside a variable.


Then an input box for the answerer, and store it inside a variable. br> 
Then a check button, and store it inside a variable.


Now put all these variables inside a single variable


Now update the HTML element who has id=’output’, with this row variable 

Now make the value null of the two input boxes



If you get stuck somewhere, you can refer to the game_page.js file from class 91, because almost everything is the same.

Submitting the Project:- If you have created project-math-quiz folder on github then -

Reupload all the files on which you have worked in the current project, in the project-math-quiz folder(this folder has been created by you in the previous project) on GitHub. You can get the steps to do this by clicking on this link.

- If you have not created project-math-quiz folder on github then -

Create a folder project-C89 and upload all the files related to the project-math-quiz project on GitHub. You can get the steps to do this by clicking on thislink.

1. Upload this HTML file on google drive. You can get the steps to do this by clicking on the following link. https://curriculum.whitehatjr.com/ADV+Asset/Project+asset/Hosting+Website.pdf
2. Copy the link which you will get after uploading your HTML file on google drive, and submit it in the Student Dashboard Projects panel against the correct class number.